requirements.txt
groq==0.24.0
openai==1.52.2
openpyxl==3.1.5
pandas==2.2.3
pip==24.2
python-dotenv==1.1.0
setuptools==68.2.0
ttkbootstrap==1.10.1
wheel==0.41.2
